+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/archiving/adapters/vpa/TableBackupAdapterVpa.java
@@ -20,7 +20,7 @@ import de.mephisto.vpin.server.popper.WheelAugmenter;
 import de.mephisto.vpin.server.system.SystemService;
 import de.mephisto.vpin.server.util.ImageUtil;
 import de.mephisto.vpin.server.util.ZipUtil;
-import de.mephisto.vpin.server.highscores.vpreg.VPReg;
+import de.mephisto.vpin.server.highscores.parsing.vpreg.VPReg;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;