+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/highscores/parsing/ScoreParsingSummary.java
@@ -1,0 +1,33 @@
+package de.mephisto.vpin.server.highscores.parsing;
+
+import de.mephisto.vpin.server.util.ScoreHelper;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class ScoreParsingSummary {
+
+  private List<ScoreParsingEntry> scores = new ArrayList<>();
+
+  public List<ScoreParsingEntry> getScores() {
+    return scores;
+  }
+
+  public void setScores(List<ScoreParsingEntry> scores) {
+    this.scores = scores;
+  }
+
+  public String toRaw() {
+    StringBuilder builder = new StringBuilder("HIGHEST SCORES\n");
+    for (ScoreParsingEntry score : scores) {
+      builder.append("#");
+      builder.append(score.getPos());
+      builder.append(" ");
+      builder.append(score.getInitials());
+      builder.append("   ");
+      builder.append(ScoreHelper.formatScore(String.valueOf(score.getScore())));
+      builder.append("\n");
+    }
+    return builder.toString();
+  }
+}