+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/highscores/parsing/text/adapters/ScoreTextFileAdapter.java
@@ -1,0 +1,6 @@
+package de.mephisto.vpin.server.highscores.parsing.text.adapters;
+
+public interface ScoreTextFileAdapter {
+
+  boolean isApplicable();
+}