+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/listeners/SubscriptionCompetitionChangeListenerImpl.java
@@ -11,7 +11,7 @@ import de.mephisto.vpin.server.discord.DiscordSubscriptionMessageFactory;
 import de.mephisto.vpin.server.games.Game;
 import de.mephisto.vpin.server.games.GameService;
 import de.mephisto.vpin.server.highscores.*;
-import de.mephisto.vpin.server.highscores.parser.HighscoreParser;
+import de.mephisto.vpin.server.highscores.parsing.HighscoreParsingService;
 import de.mephisto.vpin.server.popper.PopperService;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import org.apache.commons.lang3.StringUtils;
@@ -51,7 +51,7 @@ public class SubscriptionCompetitionChangeListenerImpl extends DefaultCompetitio
   private DiscordSubscriptionMessageFactory discordSubscriptionMessageFactory;
 
   @Autowired
-  private HighscoreParser highscoreParser;
+  private HighscoreParsingService highscoreParser;
 
   @Autowired
   private HighscoreBackupService highscoreBackupService;