+++ vpin-studio-server/src/test/java/de/mephisto/vpin/server/AbstractVPinServerTest.java
@@ -16,7 +16,7 @@ import de.mephisto.vpin.server.games.Game;
 import de.mephisto.vpin.server.games.GameDetailsRepository;
 import de.mephisto.vpin.server.games.GameEmulator;
 import de.mephisto.vpin.server.games.GameService;
-import de.mephisto.vpin.server.highscores.parser.HighscoreParser;
+import de.mephisto.vpin.server.highscores.parsing.HighscoreParsingService;
 import de.mephisto.vpin.server.highscores.HighscoreService;
 import de.mephisto.vpin.server.players.PlayerRepository;
 import de.mephisto.vpin.server.popper.PinUPConnector;
@@ -89,7 +89,7 @@ abstract public class AbstractVPinServerTest {
   protected PopperResource popperResource;
 
   @Autowired
-  protected HighscoreParser highscoreParser;
+  protected HighscoreParsingService highscoreParsingService;
 
   @Autowired
   protected PopperServiceResource popperServiceResource;